combined NewAddressModel into NewClientModal for easier conditional rendering
