+++ frontend/src/components/NewClientModal.jsx
@@ -16,7 +16,17 @@ export default function NewClientModal(props) {
     clientRateCents: 0,
   });
 
-  const [clientId, setClientId] = useState(null);
+  const [addressData, setAddressData] = useState({
+    userId: user.id,
+    line_1: '',
+    line_2: '',
+    city: '',
+    province: '',
+    country: '',
+    postalCode: '',
+  });
+
+  const [clientId, setClientId] = useState(null)
 
   const handleClientInputChange = (e) => {
     const { name, value } = e.target;
@@ -26,11 +36,19 @@ export default function NewClientModal(props) {
     }));
   };
 
+  const handleAddressInputChange = (e) => {
+    const { name, value } = e.target;
+    setAddressData((prevData) => ({
+      ...prevData,
+      [name]: value,
+    }));
+  };
+
   const handleClientModelClose = () => {
     setClientModelOpen(false);
   };
 
-  const handleSubmit = async (e) => {
+  const handleClientSubmit = async (e) => {
     e.preventDefault();
 
     try {
@@ -39,15 +57,32 @@ export default function NewClientModal(props) {
       const newClientId = response[0].id
       console.log("response", response)
       console.log(newClientId)
-      setClientId(newClientId);
+      setClientId(newClientId)
       
       // Close the modal after successful submission
-      setClientModelOpen(false);
+      // setClientModelOpen(false);
     } catch (error) {
       console.error('Error adding client:', error);
     }
   };
-   
+
+  const handleAddressSubmit = async (e) => {
+    e.preventDefault();
+
+    try {
+      // Make routes requests to send the forms data to the backend
+
+      const addressForm = {...addressData, clientId}
+      await requests.create.address(addressForm)
+      
+      
+      // Close the modal after successful submission
+      // setClientModelOpen(false);
+    } catch (error) {
+      console.error('Error adding client:', error);
+    }
+  };
+    console.log("addressData", addressData)
     console.log("clientData", clientData)
 
   return (
@@ -55,7 +90,7 @@ export default function NewClientModal(props) {
       <div className="new-client-modal-content">
         <span className="close" onClick={handleClientModelClose}>&times;</span>
         <h2>Add Client</h2>
-        <form onSubmit={handleSubmit}>
+        <form onSubmit={handleClientSubmit}>
           {/* Client Table Data*/}
           <div className="form-group">
             <label htmlFor="name">Name:</label>
@@ -114,8 +149,59 @@ export default function NewClientModal(props) {
             />
           </div>
           <button type="submit">Add Client</button>
+          </form>
+          {/* Access Table Data*/}
+          <form onSubmit={handleAddressSubmit}>
+          <div className="address-group">
+            <label>Address:</label>
+            <input
+              type="text"
+              placeholder="Line 1 Required"
+              name="line_1"
+              value={addressData.line_1}
+              onChange={handleAddressInputChange}
+              required
+            />
+            <input
+              type="text"
+              placeholder="Line 2"
+              name="line_2"
+              value={addressData.line_2}
+              onChange={handleAddressInputChange}
+            />
+            <input
+              type="text"
+              placeholder="City"
+              name="city"
+              value={addressData.city}
+              onChange={handleAddressInputChange}
+            />
+            <input
+              type="text"
+              placeholder="Province"
+              name="province"
+              value={addressData.province}
+              onChange={handleAddressInputChange}
+            />
+            <input
+              type="text"
+              placeholder="Country"
+              name="country"
+              value={addressData.country}
+              onChange={handleAddressInputChange}
+            />
+            <input
+              type="text"
+              placeholder="Postal Code Required"
+              name="postalCode"
+              value={addressData.postalCode}
+              onChange={handleAddressInputChange}
+              required
+            />
+          </div>
+          <button type="submit">Add Address</button>
         </form>
-        { <NewAddressModal setClientModelOpen={setClientModelOpen} user={user} clientId={clientId}/>}
+        {/* { <NewAddressModal setClientModelOpen={setClientModelOpen} user={user} clientId={clientId}/>} */}
       </div>
     </div>
   );
